rust_runtime: Refactor into enclave and common components
