+++ oak/server/rust/src/enclave/mod.rs
@@ -1,4 +1,3 @@
 pub mod thread;
 pub mod mutex;
-pub mod io;
 pub mod allocator;
