+++ oak/server/rust/src/lib.rs
@@ -29,53 +29,16 @@
 #![feature(array_error_internals)]
 #![feature(char_error_internals)]
 
-
 extern crate alloc;
 extern crate core;
 extern crate libc;
 
 use alloc::prelude::v1::*;
-use core::alloc::Layout;
-use core::panic::PanicInfo;
 
 pub mod enclave;
-pub mod error;
-// pub mod thread;
+pub mod common;
 
-// TODO: Move to separate crate and expose safe wrappers.
-#[link(name = "sgx_trts")]
-extern "C" {
-    // SGX-enabled abort function that causes an undefined instruction (`UD2`) to be executed, which
-    // terminates the enclave execution.
-    //
-    // The C type of this function is `extern "C" void abort(void) __attribute__(__noreturn__);`
-    //
-    // See https://github.com/intel/linux-sgx/blob/d166ff0c808e2f78d37eebf1ab614d944437eea3/sdk/trts/linux/trts_pic.S#L565.
-    fn abort() -> !;
-}
-
-#[global_allocator]
-static A: enclave::allocator::System = enclave::allocator::System;
-
-// Define what happens in an Out Of Memory (OOM) condition.
-#[alloc_error_handler]
-fn alloc_error(_layout: Layout) -> ! {
-    unsafe { abort() }
-}
-
-// See https://doc.rust-lang.org/nomicon/panic-handler.html.
-#[panic_handler]
-fn panic(_info: &PanicInfo) -> ! {
-    unsafe { abort() }
-}
-
-/// Provide the entrypoint needed by the compiler's failure mechanisms when
-/// `std` is unavailable.  See ["No
-/// stdlib" documentation](https://doc.rust-lang.org/1.2.0/book/no-stdlib.html).
-#[lang = "eh_personality"]
-pub extern "C" fn eh_personality() {}
-
-pub fn thread_test() -> enclave::io::Result<i32> {
+pub fn thread_test() -> common::io::Result<i32> {
   use alloc::sync::Arc;
   use core::sync::atomic::{AtomicI32, Ordering};
   let val = Arc::new(AtomicI32::new(2));
@@ -95,14 +58,7 @@ pub fn thread_test() -> enclave::io::Result<i32> {
 }
 
 /// An exported placeholder function to check that linking against C++ is successful.
-/// It just adds "42" to the provided value and returns it to the caller.
 #[no_mangle]
 pub extern "C" fn add_magic_number(x: i32) -> i32 {
-    // let child = thread::spawn(move || {
-    //     let v: Vec<i32> = (0..10).map(|n| n + 40).collect();
-    //     x + v[2]
-    // });
-    // let res = child.join();
-    // res
     thread_test().unwrap_or(0)
 }
